+++ benefits/src/sdk/benefitsenrollment.ts
@@ -8,8 +8,8 @@ import * as enc$ from "../lib/encodings";
 import { HTTPClient } from "../lib/http";
 import * as schemas$ from "../lib/schemas";
 import { ClientSDK, RequestOptions } from "../lib/sdks";
-import * as errors from "../sdk/models/errors";
-import * as operations from "../sdk/models/operations";
+import * as errors from "./models/errors";
+import * as operations from "./models/operations";
 
 export class BenefitsEnrollment extends ClientSDK {
     private readonly options$: SDKOptions & { hooks?: SDKHooks };
@@ -74,9 +74,13 @@ export class BenefitsEnrollment extends ClientSDK {
         } else {
             security$ = {};
         }
+        const context = {
+            operationID: "get_/benefits/enrollment/{id}",
+            oAuth2Scopes: [],
+            securitySource: this.options$.bearerAuth,
+        };
         const securitySettings$ = this.resolveGlobalSecurity(security$);
 
-        const context = { operationID: "get_/benefits/enrollment/{id}" };
         const doOptions = { context, errorCodes: [] };
         const request = this.createRequest$(
             {
@@ -139,9 +143,13 @@ export class BenefitsEnrollment extends ClientSDK {
         } else {
             security$ = {};
         }
+        const context = {
+            operationID: "get_/benefits/plan-enrollment",
+            oAuth2Scopes: [],
+            securitySource: this.options$.bearerAuth,
+        };
         const securitySettings$ = this.resolveGlobalSecurity(security$);
 
-        const context = { operationID: "get_/benefits/plan-enrollment" };
         const doOptions = { context, errorCodes: [] };
         const request = this.createRequest$(
             {
@@ -214,9 +222,13 @@ export class BenefitsEnrollment extends ClientSDK {
         } else {
             security$ = {};
         }
+        const context = {
+            operationID: "get_/benefits/plan-enrollment/{id}",
+            oAuth2Scopes: [],
+            securitySource: this.options$.bearerAuth,
+        };
         const securitySettings$ = this.resolveGlobalSecurity(security$);
 
-        const context = { operationID: "get_/benefits/plan-enrollment/{id}" };
         const doOptions = { context, errorCodes: [] };
         const request = this.createRequest$(
             {