+++ benefits/src/sdk/models/shared/planenrollment.ts
@@ -7,7 +7,7 @@ import {
     PlanEnrollmentCoverageMonth$,
 } from "./planenrollmentcoveragemonth";
 import { UserRoles, UserRoles$ } from "./userroles";
-import { z } from "zod";
+import * as z from "zod";
 
 export enum PlanType {
     MeridioHealth = "MeridioHealth",